add container files
