+++ my-app/src/actions/index.js
@@ -1,13 +1,4 @@
-
-
-
-import {FILTER_GROUNDS ,FILTER_GROUNDS_SUCCESS,FILTER_GROUNDS_ERROR,
-    SEARCH_GROUNDS,SEARCH_GROUNDS_SUCCESS,SEARCH_GROUNDS_ERROR 
-,GET_GROUNDS,GET_GROUNDS_SUCCESS ,GET_GROUNDS_ERROR ,
-SORT_GROUNDS,SORT_GROUNDS_SUCCESS,SORT_GROUNDS_ERROR,
-NAME_SORT,NAME_SORT_SUCCESS,NAME_SORT_ERROR,
-PRICE_SORT,PRICE_SORT_SUCCESS,PRICE_SORT_ERROR
-} from './styles';
+import  { FILTER_GROUNDS ,SEARCH_GROUNDS,GET_GROUNDS,SORT_GROUNDS,NAME_SORT,PRICE_SORT} from './styles';
 
 
 export const SearchGrounds = (searchValue) =>{
