+++ my-app/src/components/Comments/Comment.js
@@ -1,4 +1,4 @@
-import React from 'react'
+import  React from 'react'
 import {Link} from 'react-router-dom';
 import { Row, Col, Divider } from 'antd';
 import { Comment, Icon, Tooltip, Avatar } from 'antd';
