+++ my-app/src/index.js
@@ -17,20 +17,16 @@ import { composeWithDevTools } from 'redux-devtools-extension';
 //Middleware
 const sagaMiddleware=createSagaMiddleware();
 
+// Store
 let store =createStore(allReducers, composeWithDevTools(applyMiddleware(sagaMiddleware)));
+
+sagaMiddleware.run(rootSaga); 
 
 ReactDOM.render(
     <Provider store={store}>
         <App />
     </Provider>
 , document.getElementById('root'));
-
-// If you want your app to work offline and load faster, you can change
-// unregister() to register() below. Note this comes with some pitfalls.
-// Learn more about service workers: https://bit.ly/CRA-PWA
-
-
-sagaMiddleware.run(rootSaga); //Chạy xuyên suốt các hàm rootSaga trong app
 
 serviceWorker.unregister();
