+++ my-app/src/sagas/root.js
@@ -1,17 +1,11 @@
 // saga effect
-import { fork, all } from 'redux-saga/effects';
-import {WatchSearchGrounds,WatchGetGrounds,WatchFilterGrounds,WatchSortGrounds
-,WatchNameSort,WatchPriceSort} from './grounds';
- 
+import { all } from 'redux-saga/effects';
+import {Watch} from './grounds'; 
 
 export default function* rootSaga() {
 	 
     yield all([
-        WatchSearchGrounds(),
-        WatchGetGrounds(),
-        WatchFilterGrounds(),
-        WatchNameSort(),
-        WatchPriceSort()
+        Watch()
     ]
     )
 }
