+++ my-app/src/utils/functions.js
@@ -1,5 +1,4 @@
 
-
 // filter grounds
 export function fiterPrice(array,range) {
